Remove interim tile handling
