+++ src/ol/Tile.js
@@ -97,17 +97,8 @@ class Tile extends EventTarget {
     this.state = state;
 
     /**
-     * An "interim" tile for this tile. The interim tile may be used while this
-     * one is loading, for "smooth" transitions when changing params/dimensions
-     * on the source.
-     * @type {Tile}
-     */
-    this.interimTile = null;
-
-    /**
-     * A key assigned to the tile. This is used by the tile source to determine
-     * if this tile can effectively be used, or if a new tile should be created
-     * and this one be used as an interim tile for this new tile.
+     * A key assigned to the tile. This is used in conjunction with a source key
+     * to determine if a cached version of this tile may be used by the renderer.
      * @type {string}
      */
     this.key = '';
@@ -154,75 +145,6 @@ class Tile extends EventTarget {
    */
   getKey() {
     return this.key + '/' + this.tileCoord;
-  }
-
-  /**
-   * Get the interim tile most suitable for rendering using the chain of interim
-   * tiles. This corresponds to the  most recent tile that has been loaded, if no
-   * such tile exists, the original tile is returned.
-   * @return {!Tile} Best tile for rendering.
-   */
-  getInterimTile() {
-    if (!this.interimTile) {
-      //empty chain
-      return this;
-    }
-    let tile = this.interimTile;
-
-    // find the first loaded tile and return it. Since the chain is sorted in
-    // decreasing order of creation time, there is no need to search the remainder
-    // of the list (all those tiles correspond to older requests and will be
-    // cleaned up by refreshInterimChain)
-    do {
-      if (tile.getState() == TileState.LOADED) {
-        // Show tile immediately instead of fading it in after loading, because
-        // the interim tile is in place already
-        this.transition_ = 0;
-        return tile;
-      }
-      tile = tile.interimTile;
-    } while (tile);
-
-    // we can not find a better tile
-    return this;
-  }
-
-  /**
-   * Goes through the chain of interim tiles and discards sections of the chain
-   * that are no longer relevant.
-   */
-  refreshInterimChain() {
-    if (!this.interimTile) {
-      return;
-    }
-
-    let tile = this.interimTile;
-
-    /**
-     * @type {Tile}
-     */
-    let prev = this;
-
-    do {
-      if (tile.getState() == TileState.LOADED) {
-        //we have a loaded tile, we can discard the rest of the list
-        //we would could abort any LOADING tile request
-        //older than this tile (i.e. any LOADING tile following this entry in the chain)
-        tile.interimTile = null;
-        break;
-      } else if (tile.getState() == TileState.LOADING) {
-        //keep this LOADING tile any loaded tiles later in the chain are
-        //older than this tile, so we're still interested in the request
-        prev = tile;
-      } else if (tile.getState() == TileState.IDLE) {
-        //the head of the list is the most current tile, we don't need
-        //to start any other requests for this chain
-        prev.interimTile = tile.interimTile;
-      } else {
-        prev = tile;
-      }
-      tile = prev.interimTile;
-    } while (tile);
   }
 
   /**
